Developer randeep

+++ app/database/migrations/2021_05_05_061621_create_wallet_table.php
@@ -21,7 +21,7 @@ class CreateWalletTable extends Migration
 
             $table->boolean('status');
 
-            $table->integer('coin');
+            $table->decimal('coin',16,5);
 
             $table->unsignedBigInteger('currency_id');
